修改 DefaultMergeEventListener 实现动态更新

+++ goya-issue/goya-issue-model/src/main/java/com/goya/issue/model/po/Project.java
@@ -2,6 +2,7 @@ package com.goya.issue.model.po;
 
 import jakarta.persistence.*;
 import lombok.*;
+import org.hibernate.annotations.DynamicUpdate;
 
 import java.io.Serial;
 import java.io.Serializable;
@@ -29,7 +30,7 @@ public class Project extends BaseModel implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
 
     /**
      * name
